Hack patch to fix websocket upgrade.
I'm not even sure this should be in the project.
It feels like it should maybe be an experimental branch rather than main line.

I think the better solution would be to work out a cleaner controller interface
that allows for moving this non-core functionality out of the main project.