Move "www" to "content"
I'm split on this. I like the clarity this brings to the root of the project
but I like the simplicity of the name "www" for web content.
I can't decide if this should be www/content and www/controllers instead.

I'll give it a few weeks and see how I feel.